Refactoring partition layout code.
We need to make the partition-specific code as self-contained as
possible before we start adding code to handle a cluster layout,
otherwise things will get unmanageable.

This commit changes `chartInit()` into `drawPartition()` which handles
initial chart setup and subsequent updates. This vastly simplifies
`updateNodes()` so that we can potentially make it handle either layout
in the future.